Move `ZarrDragAndDropOpenSettings` and `ZarrOpenBehavior` from `util` to `settings`.
